移除谢肉祭的当前实现
觉得不太好，重新搞一遍

+++ kubejs/client_scripts/util/nbt.js
@@ -8,9 +8,9 @@ function ConvertPosList2Nbt(posList) {
     let res = new $ListTag()
     posList.forEach(/** @param {BlockPos} pos */pos => {
         let nbt = new $CompoundTag()
-        nbt.putInt('x', pos.getX())
-        nbt.putInt('y', pos.getY())
-        nbt.putInt('z', pos.getZ())
+        nbt.putFloat('x', pos.getX())
+        nbt.putFloat('y', pos.getY())
+        nbt.putFloat('z', pos.getZ())
         res.add(nbt)
     })
     return res
@@ -23,9 +23,9 @@ function ConvertPosList2Nbt(posList) {
  */
 function ConvertPos2Nbt(pos) {
     let nbt = new $CompoundTag()
-    nbt.putInt('x', pos.getX())
-    nbt.putInt('y', pos.getY())
-    nbt.putInt('z', pos.getZ())
+    nbt.putFloat('x', pos.getX())
+    nbt.putFloat('y', pos.getY())
+    nbt.putFloat('z', pos.getZ())
     return nbt
 }
 
@@ -38,7 +38,7 @@ function ConvertNbt2PosList(nbtList) {
     let posList = []
     nbtList.forEach(/** @param {Internal.CompoundTag} nbt */nbt => {
         if (!nbt || !nbt.contains('x') || !nbt.contains('y') || !nbt.contains('z')) return null
-        let pos = new BlockPos(nbt.getInt('x'), nbt.getInt('y'), nbt.getInt('z'))
+        let pos = new BlockPos(nbt.getFloat('x'), nbt.getFloat('y'), nbt.getFloat('z'))
         posList.push(pos)
     })
     return posList
@@ -51,6 +51,6 @@ function ConvertNbt2PosList(nbtList) {
  */
 function ConvertNbt2Pos(nbt) {
     if (!nbt || !nbt.contains('x') || !nbt.contains('y') || !nbt.contains('z')) return null
-    let pos = new BlockPos(nbt.getInt('x'), nbt.getInt('y'), nbt.getInt('z'))
+    let pos = new BlockPos(nbt.getFloat('x'), nbt.getFloat('y'), nbt.getFloat('z'))
     return pos
 }
